Move binary search to util file
